requirements.txt
# requirements.txt
langchain
langchain-community
langchain-openai
langgraph
openai
faiss-cpu
python-dotenv
tiktoken
pypdf
fastapi
uvicorn
